requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
dj-database-url==2.3.0
Django==5.1.4
django-cloudinary-storage==0.3.0
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
pillow==11.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.2
whitenoise==6.8.2
dj-rest-auth==2.1.9
djangorestframework-simplejwt

